requirements.txt
asgiref==3.7.2
Django==4.2.6
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
django-dbsettings==1.3.0